+++ models/slim.py
@@ -65,83 +65,6 @@ class LinearRecommender(BaseRecommender):
 
 
 @gin.configurable
-class UserFactorRecommender(LinearRecommender):
-    """High-dimensional matrix factorization recommender
-    
-    The model learns user factors U and item factors B
-    such that U @ B ~ X
-    where:
-    U has the shape of a user-item matrix and is regularized
-    - by constraining its sparsity pattern to that of X
-    - with a l2 penalty ~ self.l2_reg
-    B is an item-item weights matrix, regularized
-    - by constraining its diagonal to zero (as in EASE^R)
-    - by constraining its sparsity pattern to that of X.T @ X
-    - with a l2 penalty ~ weights_fn.l2_reg
-    """
-    def __init__(self, log_dir, weights_fn=None, target_density=1.0, batch_size=100, save_weights=True,
-                 als_iterations=0, l2_reg=1.0):
-
-        self.l2_reg = l2_reg
-        self.als_iterations = als_iterations
-        super().__init__(log_dir=log_dir, weights_fn=weights_fn, target_density=target_density,
-                         batch_size=batch_size, save_weights=save_weights)
-
-    def train(self, x_train, y_train, x_val, y_val):
-
-        for i in range(self.als_iterations):
-            self.weights = self.weights_fn(x_train)
-            x_train = vstack(self.user_vector(x) for x, _ in
-                             gen_batches(x_train, batch_size=1, print_interval=100))
-
-        return super().train(x_train, y_train, x_val, y_val)
-
-    def user_vector(self, x):
-        """Predict a user vector from a row x of X
-
-        Procedure
-        - B is subset to retain only rows for which x is nonzero:
-            B' = B[x > 0]
-        - we then use B' to solve a variant of the lasso:
-            argmin_u |u' @ B' - x| + l2_reg |u'|
-          where u' is similarly a subset (of u, the desired user vector)
-            u' = u[:, x > 0]
-          it has closed-form solution:
-            u' = x @ B't @ (B' @ B't + reg @ Ihh)^-1
-          with
-            Ihh = I[x > 0, x > 0]
-        - the regularization term however is tweaked to keep u close to x rather than 0
-            argmin_u |u' @ B' - x| + l2_reg |u - x'|
-          with solution
-            x @ (B'.T + reg @ Ih) @ (B' @ B't + reg @ Ihh)^-1
-          where
-            Ih = I[x > 0]
-        """
-        n_items = x.shape[1]
-        ones, hist, _ = find(x)
-        b = self.weights[hist]
-        Ihh = eye(len(hist))
-        Ihn = eye(n_items).tocsr()[hist]
-        bbt = b @ b.T + Ihh * self.l2_reg
-        xbt = x @ (b + Ihn * self.l2_reg).T
-        if issparse(xbt):
-            bbt, xbt = bbt.toarray(), xbt.toarray()
-        else:
-            bbt, xbt = np.asarray(bbt), np.asarray(xbt)
-        u = xbt @ np.linalg.inv(bbt)
-
-        return csr_matrix((u.flatten(), (ones, hist)), shape=x.shape)
-
-    def predict(self, X, y=None):
-        """Predict scores"""
-        user_vectors = vstack(self.user_vector(x) for x, _ in
-                              gen_batches(X, batch_size=1, print_interval=None))
-        y_pred = user_vectors @ self.weights
-
-        return y_pred, np.nan
-
-
-@gin.configurable
 class EmbeddingRecommender(BaseRecommender):
 
     def __init__(self, log_dir, embedding_fn=None, item_nnz=100, batch_size=100, save_embeddings=True):
